refactor code to be able to properly handle client and admin websockets
